Overhauled CV
Used the rule of no punctuation unless it's a sentence with an explicit subject or it's an imperative sentence with the implied subject of "you". Declarative sentences in bullet points that have the implied subject of "I" do not have full-stops (unless the bullet point is multiple sentences, in which case full-stops will be used to separate the sentences).

Also fixed styling of CV in order to not have any horizontal margin breaks.

Make all tables/tabular environments also not break any horizonal margins.